refactor: extract uri module from pkcs11 module
